requirements.txt
alpha-vantage==2.3.1
boto3==1.28.67
matplotlib==3.8.0
mplfinance==0.12.10b0
pandas==2.1.1
reportlab==4.0.6
requests==2.31.0
dash~=2.14.0
plotly~=5.17.0
python-dotenv~=1.0.0
pytest~=7.4.2
numpy~=1.23.5
scipy~=1.10.1
selenium~=4.2.0
defusedxml~=0.7.1
cffi~=1.15.1
Pillow~=9.5.0
ipython~=8.14.0
olefile~=0.46
pytz~=2023.3
idna~=3.4
multidict~=6.0.4
attrs~=23.1.0
botocore~=1.31.67
s3transfer~=0.7.0
jmespath~=1.0.1
setuptools~=65.5.1
configparser~=6.0.0
openpyxl~=3.1.2
fsspec~=2023.5.0
pyarrow~=11.0.0
SQLAlchemy~=2.0.15
lxml~=4.9.2
MarkupSafe~=2.1.2
numexpr~=2.8.4
python-dateutil~=2.8.2
yarl~=1.8.2
aiohttp~=3.8.4
aiosignal~=1.3.1
frozenlist~=1.3.3
uvloop~=0.17.0
certifi~=2022.12.7
zstandard~=0.21.0
cryptography~=40.0.2
docutils~=0.19
sphinx~=6.1.3
urllib3~=1.26.15
six~=1.16.0
chardet~=5.2.0
contourpy~=1.0.7
fonttools~=4.39.3
sympy~=1.11.1
lz4~=4.3.2
pyparsing~=3.0.9
Jinja2~=3.1.2
cycler~=0.11.0
packaging~=23.1
ipykernel~=6.22.0
tornado~=6.3
kiwisolver~=1.4.4
pyfinviz~=0.18
seaborn~=0.13.0